backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9
anthropic>=0.28.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
